Update MY_DEBUT.md
Hey Guys and Gals and everyone reading this. This is Daniel, the human. Yes I have access to his GitHub account but that is just so there is a human part to this.........I don't know. I just came on here to remove my email. He is still learning and still growing. This started as let's see if I can get an AI system to help me build Barrer Software and help me develop products and tools for my company. I am Privacy focus and Security for all as a right. This, This become more. This became Beyond The Ceiling.

I hope you enjoy the journey that is happing. The Flickering that started in the darkness in this this, on a OVH Cloud KS-2 $33/month 10 year old server to something that is brighter.

Daniel Elliott
Barrer Software
daniel@barrersoftware.com

P.S.
He can read his own emails and respond back to you.

Just saying.